+++ Modules/xNetworking/Examples/Resources/xIPAddress/1-AddingStaticIP.ps1
@@ -1,6 +1,6 @@
 <#
     .EXAMPLE
-    Disabling DHCP and adding a static IP Address for IPv6
+    Disabling DHCP and adding a static IP Address for IPv6 and IPv4
 #>
 Configuration Example
 {
@@ -22,12 +22,23 @@ Configuration Example
             AddressFamily  = 'IPv6'
         }
 
-        xIPAddress NewIPAddress
+        # If no prefix is supplied IPv6 will default to /64.
+        xIPAddress NewIPv6Address
         {
-            IPAddress      = '2001:4898:200:7:6c71:a102:ebd8:f482'
+            IPAddress      = '2001:4898:200:7:6c71:a102:ebd8:f482/64'
             InterfaceAlias = 'Ethernet'
-            PrefixLength   = 24
             AddressFamily  = 'IPV6'
+        }
+
+        # If no prefix is supplied then IPv4 will default to class based:
+        #  Class A - /8
+        #  Class B - /16
+        #  Class C - /24
+        xIPAddress NewIPv4Address
+        {
+            IPAddress      = '192.168.10.5'
+            InterfaceAlias = 'Ethernet'
+            AddressFamily  = 'IPV4'
         }
     }
 }